requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.2
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==1.0.1
boto3==1.35.31
botocore==1.35.31
branca==0.8.0
build==1.2.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.19.0
cohere==5.10.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.14.4
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
debugpy==1.8.6
decorator==5.1.1
deepl==1.19.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
durationpy==0.8
EbookLib==0.18
effdet==0.4.1
elastic-transport==8.15.0
elasticsearch==8.15.1
emoji==2.13.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
environs==9.5.0
et-xmlfile==1.1.0
executing==2.1.0
faiss-cpu==1.8.0.post1
fastapi==0.115.0
fastavro==1.9.7
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
FlagEmbedding==1.2.11
FlashRank==0.2.9
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
grandalf==0.8
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.23.5
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
jq==1.8.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwipiepy==0.17.1
kiwipiepy_model==0.17.0
kiwisolver==1.4.7
konlpy==0.6.0
kubernetes==31.0.0
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-chroma==0.1.4
langchain-cohere==0.1.9
langchain-community==0.2.17
langchain-core==0.2.41
langchain-elasticsearch==0.2.2
langchain-experimental==0.0.61
langchain-google-genai==1.0.10
langchain-huggingface==0.0.3
langchain-milvus==0.1.5
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langchain-upstage==0.1.7
langchainhub==0.1.21
langcodes==3.4.1
langdetect==1.0.9
langgraph==0.0.69
langsmith==0.1.129
language_data==1.2.0
lark==1.2.2
layoutparser==0.3.4
llama-index-core==0.10.68.post1
llama-index-readers-file==0.1.33
llama-parse==0.4.9
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.10
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.9.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.19.2
open_clip_torch==2.26.1
openai==1.51.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
posthog==3.6.6
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.7
PyMuPDF==1.24.10
pymupdf4llm==0.0.17
PyMuPDFb==1.24.10
pypandoc==1.13
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pysbd==0.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
python-pptx==0.6.21
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.1.20
rank-bm25==0.2.2
RapidFuzz==3.10.0
rapidocr-onnxruntime==1.3.24
redis==5.1.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.1.1
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
simsimd==5.4.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.6
streamlit==1.39.0
streamlit_folium==0.23.1
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.20.0
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typer==0.12.5
types-python-dateutil==2.9.0.20240906
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unstructured==0.10.19
unstructured-inference==0.6.6
unstructured.pytesseract==0.3.13
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.13.1
zipp==3.20.2
